+++ src/core/graphics.cc
@@ -298,12 +298,12 @@ bool IsEnabled(State cap) {
   return glIsEnabled( gl_capability[cap] );
 }
 
-void Viewport(float x, float y, float w, float h) {
+void Viewport(int32_t x, int32_t y, int32_t w, int32_t h) {
   glViewport( x, y, w, h);
 }
 
-void Viewport(float w, float h) {
-  Viewport( 0.0f, 0.0f, w, h);
+void Viewport(int32_t w, int32_t h) {
+  Viewport( 0, 0, w, h);
 }
 
 void BlendFunc(BlendFactor src_factor, BlendFactor dst_factor) {
